+++ src/renamer.rs
@@ -1,12 +1,10 @@
 use app::Config;
 use fileutils::{cleanup_files, create_backup, get_files};
-use std::collections::HashMap;
 use std::fs;
 use std::path::Path;
 use std::process;
 use std::sync::Arc;
-
-type RenameMap = HashMap<String, String>;
+use solver::{RenameMap, solve_rename_order};
 
 pub struct Renamer {
     files: Vec<String>,
@@ -36,22 +34,9 @@ impl Renamer {
             Err(_) => process::exit(1),
         };
 
-        // Return existing targets in the list of original filenames
-        let existing_targets = match existing_targets(&rename_map) {
-            Ok(existing_targets) => existing_targets,
-            Err(err) => {
-                printer.eprint(&format!(
-                    "{}Conflict with existing file: {}",
-                    colors.error.paint("Error: "),
-                    colors.error.paint(err),
-                ));
-                process::exit(1);
-            }
-        };
-
-        // Order targets to avoid renaming conflicts
-        let ordered_targets = match solve_rename_order(&rename_map, existing_targets) {
-            Ok(ordered_targets) => ordered_targets,
+        // Solve targets ordering to avoid renaming conflicts
+        let rename_order = match solve_rename_order(&rename_map) {
+            Ok(rename_order) => rename_order,
             Err(err) => {
                 printer.eprint(&format!("{}{}", colors.error.paint("Error: "), err));
                 process::exit(1);
@@ -59,7 +44,7 @@ impl Renamer {
         };
 
         // Execute actual renaming
-        for target in &ordered_targets {
+        for target in &rename_order {
             let source = &rename_map[target];
             self.rename(&source, target);
         }
@@ -162,77 +147,6 @@ impl Renamer {
             ));
         }
     }
-}
-
-/// Check if targets exists in the filesystem. If they exist, these targets must be contained in
-/// the original file list for the renaming problem to be solvable.
-fn existing_targets(rename_map: &RenameMap) -> Result<Vec<String>, String> {
-    let mut existing_targets: Vec<String> = Vec::new();
-    let sources: Vec<String> = rename_map.values().cloned().collect();
-
-    for (target, source) in rename_map {
-        if Path::new(&target).exists() {
-            if !sources.contains(&target) {
-                return Err(format!("{}->{}", source, target));
-            } else {
-                existing_targets.push(target.clone());
-            }
-        }
-    }
-    Ok(existing_targets)
-}
-
-/// Solve renaming order to avoid file overwrite. Solver will order existing targets to avoid
-/// conflicts and adds remaining targets to the list.
-fn solve_rename_order(
-    rename_map: &RenameMap,
-    mut existing_targets: Vec<String>,
-) -> Result<Vec<String>, String> {
-    // Store first all non conflicting entries
-    let mut ordered_targets: Vec<String> = rename_map
-        .iter()
-        .filter_map(|(target, _)| {
-            if !existing_targets.contains(&target) {
-                Some(target.clone())
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    // Process the container with existing targets until it is empty. The algorithm extracts
-    // recursively all targets that are not present in a container with sources exclusively related
-    // to current existing targets.
-    while !existing_targets.is_empty() {
-        // Track selected index to extract value
-        let mut selected_index: Option<usize> = None;
-        // Create a vector with all sources from existing targets
-        let sources: Vec<String> = existing_targets
-            .iter()
-            .map(|x| rename_map.get(x).cloned().unwrap())
-            .collect();
-        // Select targets without conflicts in sources
-        for (index, target) in existing_targets.iter().enumerate() {
-            if sources.contains(&target) {
-                continue;
-            } else {
-                selected_index = Some(index);
-                break;
-            }
-        }
-
-        // Store result in ordered targets container or fail to stop the loop
-        match selected_index {
-            Some(index) => ordered_targets.push(existing_targets.swap_remove(index)),
-            // This will avoid infite while loop if order is not solved
-            None => return Err("Cannot solve sorting problem.".to_string()),
-        }
-    }
-
-    // Store ordered conflicting entries
-    ordered_targets.append(&mut existing_targets);
-
-    Ok(ordered_targets)
 }
 
 #[cfg(test)]